Ensure we are authenticated during writes.
Magento writes away the session data inside a __construct().
Unfortunately this means Cm_RedisSession_Model_Session::after is called
-after- the Credis_Client is destructed.

Current behavior reconnects the client to do the write, but
authentication is not remembered.

This change moves authentication behavior to hasConnection and then
checks for it during a write. This does -not- fix the double connection
on every request, but that problem was already present. This is merely
fixing a symptom until the main issue is solved.